requirements.txt
aiogram>=3.25.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
aiosqlite>=0.20.0
yt-dlp>=2026.2.4
aiofiles>=24.1.0
Pillow>=10.0.0
pyrogram>=2.0.106
py-tgcalls>=2.2.11
TgCrypto>=1.2.5
beautifulsoup4>=4.12.0
